We fixed the issue of the scope only saving the last 1sec of a trial to workspace
The scope was only saving the last 1 second of a trial (1000 entries) when we had 10 seconds (10000 entries). To fix this we went into the following settings: Desktop Realtime Tab->prepare->control panel->configuration and signaling->trigger options->duration. This will need to be done any time we increase our time beyond 10 seconds.